Support primary selection for copying text.
This in addition to the existing support for setting the clipboard
selection.

The GLFW library unfortunately does not provide access to setting the
primary selection. So work-around it by spawning the xsel program to set the
primary selection.

If the /usr/bin/xsel program is not available, the primary selection will
not be set, but the clipboard selection still will.