Implement different color modes in mupdf-gl
Color modes are useful eg. for people using a setup primarily with
black background.

Most PDFs have black letters on white background; the new color modes
turn them into yellow letters on black background.

Mode 1 is the normal mode.

Mode 2 converts the rendered PDF image to grey-scale and maps that to
an inverted yellow-black scale, resulting in yellow text on black background.

Mode 3 is similar to mode 2 with yellow text on black background. But
it tries to preserve other colors in the PDF, by suitable color
transformation in the HSL color space.

In each new color mode, other colors (search highlight, links,
outline, etc.) are adjusted appropriately.

Color mode can be selected using the 'y' key (with number prefix to
select specific mode or without to toggle), or with -y MODE command
line option.

Color modes are implemented with OpenGL fragment shaders to remap the
colors of each pixel rendered. So they require at least OpenGL 2.0. If
shaders are not available, color mode will be disabled, with a warning
logged to the console.