Implement character-based clipboard copy
This works by dragging with the left mouse button held, similar to
most other GUI programs. Characters that are between the top and
bottom drag point, in reading order, will be selected.

This mostly relies on the order of blocks, lines, and spans in the
PDF. It does not attempt to do fancy parsing of the spacial layout of
the individual characters on a page. It seems to work well in practice
and match what other PDF viewers are doing.

Once drag selection is done, the selected characters are immediately
added to the clipboard selection. If the program xsel is available,
also the X11 primary selection will be set.

If the left mouse button is pressed over a link, a character selection
will take place if the mouse is moved before the button is released;
otherwise the link will be activated.